Move shared fields to separate file
