+++ next/schema-generator/src/generator/helpers.ts
@@ -1,8 +1,6 @@
 import type { RJSFSchema } from '@rjsf/utils'
 import camelCase from 'lodash/camelCase'
 
-import { input, object } from './functions'
-
 type ObjectJsonSchema = {
   type: 'object'
   properties: Record<string, ObjectJsonSchema | ArrayJsonSchema | RJSFSchema>
@@ -183,36 +181,6 @@ export const createCamelCaseOptionsV2 = <Option extends { title: string }>(
   return result
 }
 
-/**
- * Create phone number input field consitent with all forms.
- */
-export const inputPhoneNumber = (property: string, required: boolean, helptext?: string) =>
-  input(
-    property,
-    { title: 'Telefónne číslo', required, type: 'tel' },
-    { size: 'medium', placeholder: '+421', helptext },
-  )
-
 export const removeUndefinedValues = <T>(obj: T) => {
   return JSON.parse(JSON.stringify(obj)) as T
 }
-
-/**
- * Create address input fields.
- */
-export const inputAddress = (property: string, title: string, required: boolean) =>
-  object(property, { required }, { objectDisplay: 'boxed', title }, [
-    input('ulicaACislo', { title: 'Ulica a číslo', required, type: 'text' }, {}),
-    object(
-      'mestoPsc',
-      { required: true },
-      {
-        columns: true,
-        columnsRatio: '3/1',
-      },
-      [
-        input('mesto', { title: 'Mesto', required: true }, {}),
-        input('psc', { title: 'PSČ', required: true, format: 'zip' }, {}),
-      ],
-    ),
-  ])
